refactor: improve page dir layout
