+++ pages/artists/dedupe/index.tsx
@@ -1,9 +1,9 @@
-import { useMetaphysics } from "../../lib/artsy-next-auth"
+import { useState } from "react"
+import { useMetaphysics } from "../../../lib/artsy-next-auth"
 import {
   ArtistList,
   Skeleton,
-} from "../../components/artists/dedupe/ArtistList"
-import { useState } from "react"
+} from "../../../components/artists/dedupe/list/ArtistList"
 
 const PER_PAGE = 36
