+++ stories/ArtistList.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { ArtistList, Skeleton } from "../components/artists/dedupe/ArtistList"
+import { ArtistList, Skeleton } from "../components/artists/dedupe/list/ArtistList"
 
 export default {
   component: ArtistList,
